// apps/api/generated/server/model/registerAcceptedResponse.ts
/**
 * QZD Platform API
 * The QZD Platform API exposes authentication, account, ledger, remittance, and validator administration capabilities for wallet operators and platform integrations. The contract is defined up-front to enable client/server scaffolding via OpenAPI Generator. 
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegisterAcceptedResponse { 
    /**
     * Identifier to poll or confirm the registration flow.
     */
    registrationId: string;
    /**
     * Expiration timestamp for the registration challenge.
     */
    expiresAt: string;
}