// apps/api/generated/server/model/loginResponse.ts
/**
 * QZD Platform API
 * The QZD Platform API exposes authentication, account, ledger, remittance, and validator administration capabilities for wallet operators and platform integrations. The contract is defined up-front to enable client/server scaffolding via OpenAPI Generator. 
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LoginResponse { 
    /**
     * JWT granting access to protected routes.
     */
    accessToken: string;
    /**
     * Long lived token for session refreshes.
     */
    refreshToken: string;
    /**
     * Access token TTL in seconds.
     */
    expiresIn: number;
}